Add plane icon to map and controls to adjust mode
Mode allows selection of manual zoom/pan, following and fit to the
whole flight.

+++ src/main/java/org/flightgear/pilotlog/service/AircraftService.java
@@ -55,7 +55,9 @@ public class AircraftService {
         Aircraft summary = flightRepository.aircraftSummaryByModel(model);
         if (summary == null) {
             Aircraft aircraft = aircraftRepository.findAircraftByModel(model);
-            aircraftRepository.delete(aircraft);
+            if (aircraft != null) {
+                aircraftRepository.delete(aircraft);
+            }
         } else {
             aircraftRepository.save(summary);
         }
